requirements/test.txt
django-test-plus==1.0.18
flake8==3.3.0 # pyup: != 2.6.0
coverage==4.4.1
factory_boy==2.8.1
pytest-django==3.1.2
# pytest plugins
pytest-sugar==0.8.0
pytest-pep8==1.0.6
pytest-xdist==1.18.1
pytest-cache==0.9
pylint
Faker
pytest-factoryboy
